requirements.txt
flask
requests
yt-dlp>=2024.1.0
werkzeug
gunicorn
python-dotenv
